requirements.txt
streamlit
pypdf
sentence-transformers
langchain-groq
langchain
langchain-core
python-dotenv
rouge-score
nltk